Renamed G3D::GApp::Settings::FilmSettings -> G3D::GApp::Settings::HDRFramebufferSettings
This is most relevant to most people using the library in that you might have code that needs to be changed:

old:
app->settings().depthGuardBandThickness

settings.depthGuardBandThickness    = Vector2int16(64, 64);
settings.colorGuardBandThickness    = Vector2int16(0, 0);


new:
app->settings().hdrFramebuffer.depthGuardBandThickness
settings.hdrFramebuffer.depthGuardBandThickness    = Vector2int16(64, 64);
settings.hdrFramebuffer.colorGuardBandThickness    = Vector2int16(0, 0);

+++ G3D.lib/source/Journal.cpp
@@ -68,7 +68,7 @@ static const std::regex& headerRegex() {
 }
 
 
-static int findSection(JournalSyntax syntax, const String& fileContents, size_t start = 0) {
+static size_t findSection(JournalSyntax syntax, const String& fileContents, size_t start = 0) {
     if (syntax == JournalSyntax::DOXYGEN) {
         // Search for the section title
         size_t pos = fileContents.find("\\section", start);
